Moving functions around.
